# swagger.json
{
  "swagger": "2.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Swagger testing for Admin/Sub-Admin",
    "description": "This is assignment of Admin/sub-Admin done by vaneet kumar",
    "contact": { "email": "vaneet.k@applify.co" },
    "license": {
      "name": "",
      "url": ""
    }
  },
  "schemes": ["http", "https"],
...
  "securityDefinitions": {
    "bearerAuth": {
      "name": "Authorization",
      "in": "header",
      "type": "apiKey",
      "description": "JWT Authorization header"
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/admin/list": {
      "get": {
        "tags": ["Super-Admin/Sub-Admin/user"],
        "description": "List all the admins that are blocked or not blocked, but not deleted, alongwith count of users. And also we can filter admins by their name, if admin wants to see only blocked admins then apply blocked filter, and if admin want to see only non blocked admins then apply that filter",
        "parameters": [
          {
            "name": "limit",
            "description": "Limit",
            "in": "query",
            "required": true,
            "type": "integer",
            "value": "10"
          },
          {
            "name": "skip",
            "description": "Skip",
            "in": "query",
            "required": true,
            "type": "integer",
            "value": "0"
          },
          {
            "name": "sortBy",
            "description": "sort by",
            "in": "query",
            "required": false,
            "type": "string",
            "enum": ["name", "CreatedAt", "email", "title", "Image"],
            "value": ""
          },
          {
            "name": "orderBy",
            "description": "order by",
            "in": "query",
            "required": false,
            "type": "string",
            "enum": ["ASC", "DESC"],
            "value": ""
          },
          {
            "name": "search",
            "in": "query",
            "description": "Search",
            "type": "string",
            "value": ""
          },
          {
            "name": "Isblocked",
            "enum": [0, 1],
            "description": "Blocked admin users",
            "in": "query",
            "type": "integer",
            "required": false
          },
          {
            "name": "title",
            "enum": ["Super-Admin", "Sub-Admin", "user"],
            "description": "",
            "in": "query",
            "type": "string",
            "required": false
          }
        ],
        "responses": {}
      }
    },
...
    "/Achivement/AddAchivement": {
      "post": {
        "summary": "Add admin achivement",
        "tags": ["Admin Achivemenet"],
        "content": ["application/json"],
        "parameters": [
          {
            "name": "name",
            "in": "formData",
            "description": "Enter admin achivement",
            "required": "true",
            "type": "string"
          },
          {
            "name": "Type",
            "in": "formData",
            "description": "Enter Type",
            "required": "true",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Admin achivement Added Successfully",
            "schema": {
              "type": "object"
            }
          },
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/Achivement/list": {
      "get": {
        "tags": ["Admin Achivemenet"],
        "parameters": [
          {
            "name": "limit",
            "description": "Limit",
            "in": "query",
            "required": true,
            "type": "integer",
            "value": "10"
          },
          {
            "name": "skip",
            "description": "Skip",
            "in": "query",
            "required": true,
            "type": "integer",
            "value": "0"
          },
          {
            "name": "sortBy",
            "description": "sort by",
            "in": "query",
            "required": false,
            "type": "string",
            "enum": ["name", "CreatedAt", "Type"],
            "value": ""
          },
          {
            "name": "orderBy",
            "description": "order by",
            "in": "query",
            "required": false,
            "type": "string",
            "enum": ["ASC", "DESC"],
            "value": ""
          },
          {
            "name": "search",
            "description": "On search seen all achivement of single name and type",
            "in": "query",
            "required": false,
            "type": "string",
            "enum": [
              "Cricket-Bowling",
              "Cricket-Batsman",
              "Stage",
              "Sequential",
              "Parallel",
              "Sequential"
            ],
            "value": ""
          }
        ],
        "responses": {}
      }
    },
    "/Achivement/edit": {
      "put": {
        "summary": "Edit admin Achivemenet.Need to login first and need token",
        "tags": ["Admin Achivemenet"],
        "content": ["application/json"],
        "parameters": [
          {
            "name": "Id",
            "in": "formData",
            "description": "Enter Id",
            "required": "true",
            "type": "string"
          },
          {
            "name": "name",
            "in": "formData",
            "description": "",
            "required": "true",
            "type": "string"
          },
          {
            "name": "Type",
            "in": "formData",
            "description": "",
            "required": "true",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Edit Successfull",
            "schema": {
              "type": ""
            }
          },
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/Achivement/delete": {
      "delete": {
        "tags": ["Admin Achivemenet"],
        "content": ["application/json"],
        "parameters": [
          {
            "name": "Id",
            "in": "formData",
            "description": "Enter Admin achivement Id",
            "required": "true",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Delete Successfull",
            "schema": {
              "type": ""
            }
          },
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
...
    "/api/delete/{id}": {
      "delete": {
        "summary": "Delete user.Need to login first and need token",
        "tags": ["User"],
        "content": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "formData",
            "description": "Enter User Id",
            "required": "true",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Delete Successfull",
            "schema": {
              "type": ""
            }
          },
          "405": {
            "description": "Invalid input"
          }
        }
      }
    }
  }
}